split up math utilities into dynamics, kinematics and geometry, add zone consciousness
