+++ lance/src/robot/control/shared/mining_controller.cpp
@@ -41,8 +41,8 @@
 
 #include <limits>
 
-#include "robot/core/robot_math.hpp"
 #include "robot/core/hid_bindings.hpp"
+#include "robot/model/dynamics.hpp"
 
 
 namespace lance
