+++ lance/src/robot/model/geometry.hpp
@@ -1,0 +1,194 @@
+/*******************************************************************************
+*   Copyright (C) 2025-2026 Cardinal Space Mining Club                         *
+*                                                                              *
+*                                 ;xxxxxxx:                                    *
+*                                ;$$$$$$$$$       ...::..                      *
+*                                $$$$$$$$$$x   .:::::::::::..                  *
+*                             x$$$$$$$$$$$$$$::::::::::::::::.                 *
+*                         :$$$$$&X;      .xX:::::::::::::.::...                *
+*                 .$$Xx++$$$$+  :::.     :;:   .::::::.  ....  :               *
+*                :$$$$$$$$$  ;:      ;xXXXXXXXx  .::.  .::::. .:.              *
+*               :$$$$$$$$: ;      ;xXXXXXXXXXXXXx: ..::::::  .::.              *
+*              ;$$$$$$$$ ::   :;XXXXXXXXXXXXXXXXXX+ .::::.  .:::               *
+*               X$$$$$X : +XXXXXXXXXXXXXXXXXXXXXXXX; .::  .::::.               *
+*                .$$$$ :xXXXXXXXXXXXXXXXXXXXXXXXXXXX.   .:::::.                *
+*                 X$$X XXXXXXXXXXXXXXXXXXXXXXXXXXXXx:  .::::.                  *
+*                 $$$:.XXXXXXXXXXXXXXXXXXXXXXXXXXX  ;; ..:.                    *
+*                 $$& :XXXXXXXXXXXXXXXXXXXXXXXX;  +XX; X$$;                    *
+*                 $$$: XXXXXXXXXXXXXXXXXXXXXX; :XXXXX; X$$;                    *
+*                 X$$X XXXXXXXXXXXXXXXXXXX; .+XXXXXXX; $$$                     *
+*                 $$$$ ;XXXXXXXXXXXXXXX+  +XXXXXXXXx+ X$$$+                    *
+*               x$$$$$X ;XXXXXXXXXXX+ :xXXXXXXXX+   .;$$$$$$                   *
+*              +$$$$$$$$ ;XXXXXXx;;+XXXXXXXXX+    : +$$$$$$$$                  *
+*               +$$$$$$$$: xXXXXXXXXXXXXXX+      ; X$$$$$$$$                   *
+*                :$$$$$$$$$. +XXXXXXXXX;      ;: x$$$$$$$$$                    *
+*                ;x$$$$XX$$$$+ .;+X+      :;: :$$$$$xX$$$X                     *
+*               ;;;;;;;;;;X$$$$$$$+      :X$$$$$$&.                            *
+*               ;;;;;;;:;;;;;x$$$$$$$$$$$$$$$$x.                               *
+*               :;;;;;;;;;;;;.  :$$$$$$$$$$X                                   *
+*                .;;;;;;;;:;;    +$$$$$$$$$                                    *
+*                  .;;;;;;.       X$$$$$$$:                                    *
+*                                                                              *
+*   Unless required by applicable law or agreed to in writing, software        *
+*   distributed under the License is distributed on an "AS IS" BASIS,          *
+*   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.   *
+*   See the License for the specific language governing permissions and        *
+*   limitations under the License.                                             *
+*                                                                              *
+*******************************************************************************/
+
+#pragma once
+
+#include <cmath>
+#include <limits>
+
+#include <Eigen/Core>
+#include <Eigen/Geometry>
+
+#include "util/geometry.hpp"
+
+
+#ifndef LANCE
+    #define LANCE 2
+#endif
+
+#define CONSTEXPR_VAL_TEMPLATE(name, val)             \
+    template<typename T>                              \
+    inline constexpr T name##_ = static_cast<T>(val); \
+    inline constexpr double name = name##_<double>;
+
+namespace lance
+{
+
+namespace geom
+{
+
+#if LANCE <= 1
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_X_MAX, 0.735)
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_X_MIN, -0.765)
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_Y_MAX, 0.369)
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_Y_MIN, -0.369)
+CONSTEXPR_VAL_TEMPLATE(COLLISION_Z_MAX, 0.675)
+CONSTEXPR_VAL_TEMPLATE(COLLISION_Z_MIN, -0.102)
+#elif LANCE >= 2
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_X_MAX, 0.591)
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_X_MIN, -0.640)
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_Y_MAX, 0.362)
+CONSTEXPR_VAL_TEMPLATE(FOOTPRINT_Y_MIN, -0.362)
+CONSTEXPR_VAL_TEMPLATE(COLLISION_Z_MAX, 0.810)
+CONSTEXPR_VAL_TEMPLATE(COLLISION_Z_MIN, -0.102)
+#endif
+
+CONSTEXPR_VAL_TEMPLATE(PRIMARY_COLLISION_ZONE_X, FOOTPRINT_X_MIN)
+CONSTEXPR_VAL_TEMPLATE(
+    PRIMARY_COLLISION_ZONE_Y,
+    ((FOOTPRINT_Y_MAX + FOOTPRINT_Y_MIN) / 2))
+CONSTEXPR_VAL_TEMPLATE(
+    PRIMARY_COLLISION_ZONE_Z,
+    ((COLLISION_Z_MAX + COLLISION_Z_MIN) / 2))
+
+CONSTEXPR_VAL_TEMPLATE(
+    PRIMARY_COLLISION_ZONE_LENGTH_OFFSET,
+    (FOOTPRINT_X_MAX - FOOTPRINT_X_MIN))
+CONSTEXPR_VAL_TEMPLATE(
+    PRIMARY_COLLISION_ZONE_WIDTH,
+    (FOOTPRINT_Y_MAX - FOOTPRINT_Y_MIN))
+CONSTEXPR_VAL_TEMPLATE(
+    PRIMARY_COLLISION_ZONE_HEIGHT,
+    (COLLISION_Z_MAX - COLLISION_Z_MIN))
+
+
+
+template<typename T>
+using Vec2 = Eigen::Vector2<T>;
+template<typename T>
+using Vec3 = Eigen::Vector3<T>;
+template<typename T>
+using Pose2 = Vec3<T>;
+template<typename T>
+
+using Quat = Eigen::Quaternion<T>;
+template<typename T>
+using Iso3 = Eigen::Transform<T, 3, Eigen::Isometry>;
+template<typename T>
+using Pose3 = util::geom::Pose3<T>;
+template<typename T>
+using PoseTf3 = util::geom::PoseTf3<T>;
+template<typename T>
+
+using Box2 = Eigen::AlignedBox<T, 2>;
+template<typename T>
+using Box3 = Eigen::AlignedBox<T, 3>;
+
+
+using Vec2f = Vec2<float>;
+using Vec3f = Vec3<float>;
+using Pose2f = Pose2<float>;
+
+using Quatf = Quat<float>;
+using Iso3f = Iso3<float>;
+using Pose3f = util::geom::Pose3f;
+using PoseTf3f = util::geom::PoseTf3f;
+
+using Box2f = Eigen::AlignedBox2f;
+using Box3f = Eigen::AlignedBox3f;
+
+
+template<typename T>
+constexpr inline T quatToYaw(const Quat<T>& q)
+{
+    return std::atan2(
+        (2 * (q.w() * q.z() + q.x() * q.y())),
+        (1 - 2 * (q.y() * q.y() + q.z() * q.z())));
+}
+template<typename T>
+constexpr inline Pose2<T> flattenPose(const Pose3<T>& p)
+{
+    return Pose2<T>{p.vec.x(), p.vec.y(), quatToYaw(p.quat)};
+}
+
+/* Obtain the headroom that the robot could travel before reaching the zone 
+ * boundary. Note that the calculation bases of the FRONT of the robot and not
+ * the center point. */
+template<typename T>
+constexpr inline T distToBounds(const Pose2<T>& p, const Box2<T>& b)
+{
+    if(b.contains(p.template head<2>()))
+    {
+        const T dx = std::cos(p.z());
+        const T dy = std::sin(p.z());
+        T tx = std::numeric_limits<T>::max();
+        T ty = std::numeric_limits<T>::max();
+
+        if(std::abs(dx) > 1e-6)
+        {
+            tx = ((dx > 0 ? b.max().x() : b.min().x()) - p.x()) / dx;
+        }
+        if(std::abs(dy) > 1e-6)
+        {
+            ty = ((dy > 0 ? b.max().y() : b.min().y()) - p.y()) / dy;
+        }
+
+        return std::min(tx, ty) - FOOTPRINT_X_MAX_<T>;
+    }
+    return std::numeric_limits<T>::max();
+}
+/* Flattends 3d pose to 2d and applies distToBounds() overload for Pose2. */
+template<typename T>
+constexpr inline T distToBounds(const Pose3<T>& p, const Box2<T>& b)
+{
+    return distToBounds(flattenPose(p), b);
+}
+/* Flattends 3d pose to 2d and applies distToBounds() overload for Pose2. */
+template<typename T>
+constexpr inline T distToBounds(const PoseTf3<T>& p, const Box2<T>& b)
+{
+    return distToBounds(p.pose, b);
+}
+
+
+};  // namespace geom
+
+};  // namespace lance
+
+#undef CONSTEXPR_VAL_TEMPLATE
